make copying better

+++ copying.go
@@ -12,6 +12,13 @@ import (
 	"time"
 )
 
+var (
+	newFileName string
+	// If the folder contains more than one .pak file (e.g. Stardust Core has two .pak files
+	// in its folder), a prefix will be added to the result files
+	prefix int
+)
+
 // Check if the target folder exist and create one if not
 func checkTargetDirectory(targetDir string) {
 	if _, err := os.Stat(targetDir); errors.Is(err, os.ErrNotExist) {
@@ -53,12 +60,14 @@ func copyFiles(d directories, scrapTitles bool, wg *sync.WaitGroup) {
 		// going through all the directories of source folder
 		for _, entry := range entries {
 			if entry.IsDir() {
-				// save the name (steam id)
+				// reset the prefix
+				prefix = 1
+				// save the name (Steam id)
 				currentDirectoryName := entry.Name()
 				// path to the current directory
 				folderPath := filepath.Join(*d.srcDirectoryPath, currentDirectoryName)
 
-				// all entries (files) of directory with content
+				// all entries (files) of directory with content (.pak files expectedly)
 				files, err := os.ReadDir(folderPath)
 				if err != nil {
 					panic(err)
@@ -69,19 +78,26 @@ func copyFiles(d directories, scrapTitles bool, wg *sync.WaitGroup) {
 					if !file.IsDir() {
 						// check if the current file is ".pak"
 						if filepath.Ext(file.Name()) == ".pak" {
+							// get the path of .pak file from Starbound's workshop folder
 							currentFileName := file.Name()
 							currentFilePath := filepath.Join(folderPath, currentFileName)
 
-							newFileName := getWorkshopItemTitle(currentDirectoryName) + ".pak"
+							// set new file name's to a title in the Steam workshop
+							newFileName = getWorkshopItemTitle(currentDirectoryName) + ".pak"
+							if prefix > 1 {
+								newFileName = getWorkshopItemTitle(currentDirectoryName) + "_" + fmt.Sprint(prefix) + ".pak"
+							}
+							prefix++
+
+							// path to a new file with its name and target folder
 							newFilePath := filepath.Join(*d.targetDirectoryPath, newFileName)
 
+							// skip if the file already exists
 							if _, err := os.Stat(newFilePath); errors.Is(err, os.ErrNotExist) {
 								err := copyFile(currentFilePath, newFilePath)
 								if err != nil {
 									panic(err)
 								}
-							} else {
-								fmt.Println("skipped")
 							}
 						}
 					}
@@ -92,12 +108,14 @@ func copyFiles(d directories, scrapTitles bool, wg *sync.WaitGroup) {
 		// going through all the directories of source folder
 		for _, entry := range entries {
 			if entry.IsDir() {
-				// save the name (steam id)
+				// reset the prefix
+				prefix = 1
+				// save the name (Steam id)
 				currentDirectoryName := entry.Name()
 				// path to the current directory
 				folderPath := filepath.Join(*d.srcDirectoryPath, currentDirectoryName)
 
-				// all entries (files) of directory with content
+				// all entries (files) of directory with content (.pak files expectedly)
 				files, err := os.ReadDir(folderPath)
 				if err != nil {
 					panic(err)
@@ -108,15 +126,26 @@ func copyFiles(d directories, scrapTitles bool, wg *sync.WaitGroup) {
 					if !file.IsDir() {
 						// check if the current file is ".pak"
 						if strings.Split(file.Name(), ".")[1] == "pak" {
+							// get the path of .pak file from Starbound's workshop folder
 							currentFileName := file.Name()
 							currentFilePath := filepath.Join(folderPath, currentFileName)
 
-							newFileName := currentDirectoryName + ".pak"
+							// set new file name's to an ID in the Steam workshop
+							newFileName = currentDirectoryName + ".pak"
+							if prefix > 1 {
+								newFileName = currentDirectoryName + "_" + fmt.Sprint(prefix) + ".pak"
+							}
+							prefix++
+
+							// path to a new file with its name and target folder
 							newFilePath := filepath.Join(*d.targetDirectoryPath, newFileName)
 
-							err := copyFile(currentFilePath, newFilePath)
-							if err != nil {
-								panic(err)
+							// skip if the file already exists
+							if _, err := os.Stat(newFilePath); errors.Is(err, os.ErrNotExist) {
+								err := copyFile(currentFilePath, newFilePath)
+								if err != nil {
+									panic(err)
+								}
 							}
 						}
 					}
